Pagination and cards access

+++ app/routes/cards/cardsCRUD.js
@@ -1,27 +1,34 @@
 const createModelMiddleware = require('../middleware/createModelMiddleware');
+const checkEndpointAccessMiddleware = require('../../auth/middleware/checkEndpointAccessMiddleware');
 const findModelByMiddleware = require('../middleware/findModelByMiddleware');
+const paginateMiddleware = require('../middleware/paginateMiddleware');
 const {CardModel} = require('../../models/appModels');
 
 module.exports = (router) => {
 
     router.get('/cards',
+        checkEndpointAccessMiddleware(4),
+        paginateMiddleware(CardModel),
         async (req, res) => {
-            res.jsonSuccess(await CardModel.find({}).exec());
+
         });
 
     router.get('/cards/:id',
+        checkEndpointAccessMiddleware(2),
         findModelByMiddleware(CardModel, {_id: 'id'}, 'params', 'card'),
         async (req, res) => {
             res.jsonSuccess(req.card);
         });
 
     router.post('/cards',
+        checkEndpointAccessMiddleware(4),
         createModelMiddleware(CardModel, 'card'),
         (req, res) => {
             res.jsonSuccess(req.card);
         });
 
     router.patch('/cards/:id',
+        checkEndpointAccessMiddleware(4),
         findModelByMiddleware(CardModel, {_id: 'id'}, 'params', 'card'),
         async (req, res) => {
             try {
@@ -36,7 +43,8 @@ module.exports = (router) => {
         });
 
     router.delete('/cards/:id',
-        findModelByMiddleware(CardModel, {_id: 'id'}, 'params', 'card'),
+        checkEndpointAccessMiddleware(4),
+        findModelByMiddleware(CardModel, {_id: 'id'}, 'req.params', 'card'),
         async (req, res) => {
             await req.card.delete();
             res.jsonSuccess();
